Migración catálogo a SQL, litros como decimal y seguridad
- El catálogo de productos ahora se carga desde SQL, eliminando el CSV.
- Unificación de litros/cantidades a tipo decimal en todos los modelos.
- El Worker enriquece ventas/históricos con datos del catálogo.
- Dashboard ahora grafica litros y soporta periodos flexibles ("Hoy", "Mes", "Año").
- Implementado hash seguro de contraseñas (PBKDF2) en AuthService.
- Refactorización para usar interfaces y DI, cadenas de conexión por config.
- Mejoras visuales: periodo "Hoy" por defecto, etiquetas más claras.
- Limpieza y regeneración de archivos de recursos/settings.
- Nuevas dependencias: ConfigurationManager y DependencyInjection.

+++ MVVM/ViewModels/FamiliaViewModel.cs
@@ -299,7 +299,7 @@ namespace WPF_PAR.MVVM.ViewModels
                 _datosFamiliaActual = ventasCompletas.Where(x => x.Familia == familia).ToList();
 
                 GranTotalVenta = _datosFamiliaActual.Sum(x => x.TotalVenta);
-                GranTotalLitros = _datosFamiliaActual.Sum(x => x.LitrosTotales);
+                GranTotalLitros = (double)_datosFamiliaActual.Sum(x => x.LitrosTotales);
                 TituloReporteCard = "📥 DESCARGAR DETALLE";
 
                 OnPropertyChanged(nameof(GranTotalVenta));
@@ -369,7 +369,7 @@ namespace WPF_PAR.MVVM.ViewModels
 
             SeriesPastelLitros = resumenDatos.OrderByDescending(x => x.Litros).Take(5).Select(x => new PieSeries<double>
             {
-                Values = new double[] { x.Litros },
+                Values = new double[] { (double)x.Litros },
                 Name = NormalizarNombreProducto(x.Nombre),
                 InnerRadius = 0,
                 DataLabelsFormatter = p => $"{p.Model:N0} L ({p.StackedValue.Share:P0})",
@@ -384,7 +384,7 @@ namespace WPF_PAR.MVVM.ViewModels
             {
                 Descripcion = string.IsNullOrEmpty(x.Cliente) ? "PÚBLICO GENERAL" : x.Cliente,
                 TotalVenta = x.TotalVenta,
-                LitrosTotal = x.LitrosTotales // <--- ChartService probablemente usa esta para gráficas
+                LitrosTotal = (double)x.LitrosTotales // <--- ChartService probablemente usa esta para gráficas
             }).ToList();
 
             var resClientes = _chartService.GenerarTopProductos(datosParaClientes, VerPorLitros, TopSeleccionado);
