+++ WPF_PAR.Core/Models/VentaReporteModel.cs
@@ -16,12 +16,12 @@ namespace WPF_PAR.Core.Models
         public string Familia { get; set; }
         public string Linea { get; set; }
         public string Color { get; set; }
-        public double Cantidad { get; set; }
+        public decimal Cantidad { get; set; }
         public decimal PrecioUnitario { get; set; }
         public decimal Descuento { get; set; }
         public decimal TotalVenta { get; set; }
-        public double LitrosUnitarios { get; set; }
-        public double LitrosTotales => ( double ) Cantidad * LitrosUnitarios;
+        public decimal LitrosUnitarios { get; set; }
+        public decimal LitrosTotales => Cantidad * LitrosUnitarios;
         public double LitrosTotal { get; set; }
 
         public string Agente { get; set; }
